Level Design 3/27
I fixed some previous issues, and added in a new asset pack to give our floor some actual color. The previous pack is being weird. This is NOT my final thing I'll be doing for level design. I want to add more and give actual set design, but I wanted to add this just for now.